5. End To End Document Q&A Using Google Gemma, Groq API/requirements.txt
streamlit
python-dotenv
pillow
langchain
PyPDF2  # reading pdfs
faiss-cpu # or use -> chromadb # for vector embeddings
langchain_google_genai
groq
langchain-groq